LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/reporterror.any-expected.txt

CONSOLE MESSAGE: 1
CONSOLE MESSAGE: TypeError
CONSOLE MESSAGE: undefined

PASS self.reportError(1)
PASS self.reportError(TypeError)
PASS self.reportError(undefined)
PASS self.reportError() (without arguments) throws
